refactor: adjust structure
