+++ dmconvert/guardgift/gg_handler.py
@@ -1,6 +1,24 @@
 # Copyright (c) 2025 DanmakuConvert
 
-from .guard_and_gift import extract_gift_data, merge_gifts, adjust_time_conflicts, calculate_moves, generate_ass_line
+from .guard_and_gift import merge_gifts, adjust_time_conflicts, calculate_moves, generate_ass_line
+
+def extract_gift_data(element):
+    """extract the common attributes of gifts and guards"""
+    fixed_time = 2 # the time of gift danmaku
+    data = {
+        'appear_time': float(element.get('ts')),
+        'over_time': float(element.get('ts')) + fixed_time,  
+        'user': element.get('user'),
+        'name': element.get('giftname'),
+        'count': int(element.get('giftcount' if element.tag=='gift' else 'count')), 
+        'price': element.get('price'),
+        'move': 0,
+        'height': 0,
+        'move_time': -1,
+        'disappear_time': -2,
+        'isdis': 0 
+    }
+    return data
 
 def draw_gift_and_guard(ass_file, root, font_size, resolution_y):
     with open(ass_file, 'a', encoding='utf-8') as f:
